Adding docs on PCR values and experimentation script
I'm adding a bunch of documents covering the research we did on PCRs,
which events extend each of them, what each of those events measure and
in which order. I'm adding some conclusions as well as highlighting the
main gap: verifying the kernel command line on a GRUB system.

The entry-point is the `pcrs/grub.md` file. The rest of the files added
in this commit are referenced from there. It contains a little breakdown
of the events of each PCR, as well as a little experiment that boots a
FCOS VM with a TPM and OVMF (UEFI) firmware, and the eventlog resulting
from it.

I'm not documenting the PCR1 events as they are related to OVMF
configuration. FCOS can be deployed on confidential VMs using OVMF.
However, we are limiting the PCRs to those that are filled by the
FCOS side, rather than the firmware side, as they can be extended to any
platform, no matter the firmware. Firmware, however, will change from
one platform to another, which we could support at some point, but it is
not the main goal at the moment.

I'm not documenting PCR2 and PCR3 either, for the same reason as above.

I'm not sure about how predictable the GPT is, or the benefit we could
obtain from it.

I'm documenting the conclusions that you guys reached some time ago, on
which PCRs to look at for attestation. I'm also documenting the main
gap, which is not being able to predict the value of PCR8 at the moment,
and so, not being able to verify the kernel command line is fine.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>