Fix the stuck issue about get ip
The 'populate-trustee-kbs.sh' script would fail if the trustee VM had not yet obtained an IP address when the script was initiated. This resulted in an infinite loop while waiting for the KBS. This change moves the IP address retrieval inside the 'until' loop, ensuring the script repeatedly attempts to get the IP until itsucceeds.

Signed-off-by: Yalan Zhang <yalzhang@redhat.com>